[refactor] LottoMatching(model) - 당첨번호 매칭
